v3.2.7 - only for items now might investigate for models but not now
