--- src/main/java/net/enchantoutline/Shaders.java
@@ -1,93 +1,0 @@
-package net.enchantoutline;
-
-import com.mojang.blaze3d.pipeline.RenderTarget;
-import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.DefaultVertexFormat;
-import com.mojang.blaze3d.vertex.VertexFormat;
-
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.GameRenderer;
-import net.minecraft.client.renderer.RenderStateShard;
-import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.ShaderInstance;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.inventory.InventoryMenu;
-import org.lwjgl.opengl.GL11;
-
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-public class Shaders {
-
-    public static ShaderInstance itemShaderInstance;
-    public static ShaderInstance entityShaderInstance;
-    private static RenderType itemLayer;
-    private static final Map<ResourceLocation, RenderType> entityLayerCache = new ConcurrentHashMap<>();
-    private static final Map<ResourceLocation, RenderType> testCache = new ConcurrentHashMap<>();
-
-    public static RenderType getItemOutlineLayer() {
-        if (itemLayer == null) itemLayer = RenderType.create(
-                "item",
-                DefaultVertexFormat.NEW_ENTITY,
-                VertexFormat.Mode.QUADS,
-                1536,
-                false, false,
-                RenderType.CompositeState.builder()
-                        .setShaderState(new RenderStateShard.ShaderStateShard(() -> {
-                            if (itemShaderInstance != null) return itemShaderInstance;
-                            return GameRenderer.getPositionTexShader();
-                        }))
-                        .setTextureState(new RenderStateShard.TextureStateShard(InventoryMenu.BLOCK_ATLAS, false, false))
-                        .setWriteMaskState(RenderStateShard.COLOR_WRITE)
-                        .setCullState(RenderStateShard.NO_CULL)
-                        .setDepthTestState(RenderStateShard.LEQUAL_DEPTH_TEST)
-                        .setTransparencyState(RenderStateShard.TRANSLUCENT_TRANSPARENCY)
-                        .setLayeringState(RenderStateShard.VIEW_OFFSET_Z_LAYERING)
-                        .createCompositeState(false)
-        );
-
-        return itemLayer;
-    }
-
-    public static RenderType getEntityOutlineLayer(ResourceLocation location) {
-        return entityLayerCache.computeIfAbsent(location, texture -> RenderType.create(
-                "armor",
-                DefaultVertexFormat.NEW_ENTITY,
-                VertexFormat.Mode.QUADS,
-                1536,
-                false,
-                true,
-                RenderType.CompositeState.builder()
-                        .setShaderState(new RenderStateShard.ShaderStateShard(() -> entityShaderInstance))
-                        .setTextureState(new RenderStateShard.TextureStateShard(texture, false, false))
-                        .setTransparencyState(RenderStateShard.TRANSLUCENT_TRANSPARENCY)
-                        .setLightmapState(RenderStateShard.LIGHTMAP)
-                        .setOverlayState(RenderStateShard.OVERLAY)
-                        .setDepthTestState(RenderStateShard.LEQUAL_DEPTH_TEST)
-                        .setCullState(RenderStateShard.NO_CULL)
-                        .setWriteMaskState(RenderStateShard.COLOR_WRITE)
-                        .createCompositeState(false)
-        ));
-    }
-
-    public static RenderType test(ResourceLocation location) {
-        return testCache.computeIfAbsent(location, texture -> RenderType.create(
-                "test",
-                DefaultVertexFormat.NEW_ENTITY,
-                VertexFormat.Mode.QUADS,
-                1536,
-                false,
-                false,
-                RenderType.CompositeState.builder()
-                        .setShaderState(RenderStateShard.RENDERTYPE_ENTITY_SOLID_SHADER)
-                        .setTextureState(new RenderStateShard.TextureStateShard(location, false, false))
-                        .setTransparencyState(RenderStateShard.TRANSLUCENT_TRANSPARENCY)
-                        .setLightmapState(RenderStateShard.LIGHTMAP)
-                        .setOverlayState(RenderStateShard.OVERLAY)
-                        .setDepthTestState(RenderStateShard.LEQUAL_DEPTH_TEST) // Test depth (don't show through walls)
-                        .setCullState(RenderStateShard.NO_CULL)
-                        .setWriteMaskState(RenderStateShard.COLOR_WRITE)
-                        .createCompositeState(false)
-        ));
-    }
-}
