--- src/main/java/net/enchantoutline/mixin/CustomHeadLayerMixin.java
@@ -1,8 +1,0 @@
-package net.enchantoutline.mixin;
-
-import net.minecraft.client.renderer.entity.layers.CustomHeadLayer;
-import org.spongepowered.asm.mixin.Mixin;
-
-@Mixin(CustomHeadLayer.class)
-public class CustomHeadLayerMixin {
-}
