--- src/main/java/net/enchantoutline/mixin/ElytraLayerMixin.java
@@ -1,8 +1,0 @@
-package net.enchantoutline.mixin;
-
-import net.minecraft.client.renderer.entity.layers.ElytraLayer;
-import org.spongepowered.asm.mixin.Mixin;
-
-@Mixin(ElytraLayer.class)
-public class ElytraLayerMixin {
-}
