--- src/main/resources/assets/enchant_outline/shaders/core/entity.fsh
@@ -1,24 +1,0 @@
-#version 150
-
-uniform sampler2D Sampler0;
-uniform vec4 ColorModulator;
-uniform vec4 GlowColor;
-
-in vec2 texCoord0;
-in vec4 vertexColor;
-
-out vec4 fragColor;
-
-void main() {
-    vec4 textureColor = texture(Sampler0, texCoord0);
-    if (textureColor.a < 0.1) {
-        discard;
-    }
-
-    if (gl_FrontFacing) {
-        discard;
-    }
-
-    fragColor = GlowColor * ColorModulator; // no lightmap
-    //fragColor = GlowColor * ColorModulator * vertexColor; // with lightmap (test)
-}
