--- src/main/resources/assets/enchant_outline/shaders/core/entity.json
@@ -1,31 +1,0 @@
-{
-    "blend": {
-        "func": "add",
-        "srcrgb": "srcalpha",
-        "dstrgb": "1-srcalpha",
-        "funcalpha": "add",
-        "srcalpha": "one",
-        "dstalpha": "zero"
-    },
-    "vertex": "enchant_outline:entity",
-    "fragment": "enchant_outline:entity",
-    "attributes": [
-        "Position",
-        "Color",
-        "UV0",
-        "UV1",
-        "UV2",
-        "Normal"
-    ],
-    "uniforms": [
-        { "name": "ModelViewMat",  "type": "matrix4x4", "count": 16, "values": [ 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0 ] },
-        { "name": "ProjMat",       "type": "matrix4x4", "count": 16, "values": [ 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0 ] },
-        { "name": "TextureMat",    "type": "matrix4x4", "count": 16, "values": [ 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0 ] },
-        { "name": "ColorModulator","type": "float",     "count": 4,  "values": [ 1.0, 1.0, 1.0, 1.0 ] },
-        { "name": "Thickness",     "type": "float",     "count": 1,  "values": [ 0.01 ] },
-        { "name": "GlowColor",     "type": "float",     "count": 4,  "values": [ 0.85, 0.70, 0.25, 1.00 ] }
-    ],
-    "samplers": [
-        { "name": "Sampler0" }
-    ]
-}
