+++ src/main/resources/assets/enchant_outline/shaders/core/item.json
@@ -13,7 +13,6 @@
         "Position",
         "Color",
         "UV0",
-        "UV1",
         "UV2",
         "Normal"
     ],
