+++ src/main/resources/assets/enchant_outline/shaders/core/test.fsh
@@ -1,0 +1,26 @@
+#version 150
+
+#moj_import <minecraft:fog.glsl>
+
+uniform sampler2D Sampler0;
+
+uniform vec4 ColorModulator;
+uniform float FogStart;
+uniform float FogEnd;
+uniform vec4 FogColor;
+
+in float vertexDistance;
+in vec4 vertexColor;
+in vec2 texCoord0;
+
+out vec4 fragColor;
+
+void main() {
+    vec4 color = texture(Sampler0, texCoord0) * vertexColor * ColorModulator;
+    #ifdef ALPHA_CUTOUT
+    if (color.a < ALPHA_CUTOUT) {
+        discard;
+    }
+    #endif
+    fragColor = linear_fog(vertexColor, vertexDistance, FogStart, FogEnd, FogColor);
+}
