+++ src/main/resources/enchant_outline.mixins.json
@@ -10,9 +10,6 @@
   },
   "client": [
     "BlockEntityWithoutLevelRendererMixin",
-    "CustomHeadLayerMixin",
-    "ElytraLayerMixin",
-    "HumanoidArmorLayerMixin",
     "ItemRendererMixin",
     "ThrownTridentRendererMixin"
   ]
